+++ CHANGELOG.rst
@@ -1,7 +1,12 @@
+1.2.5
+=====
+
+* Added compatibility with Arango 2.8.x
+
 1.2.4
 =====
 
-* missing import in collections.py added
+* Missing import in collections.py added
 
 1.2.3
 =====